+++ lib/statechart.ex
@@ -5,23 +5,108 @@ defmodule Statechart do
   @external_resource "README.md"
   @moduledoc usage_section
 
-  alias Statechart.Build.MacroOnEnterExit
+  alias Statechart.Build.AccStep
   alias Statechart.Build.MacroState
-  alias Statechart.Build.MacroStatechart
+  alias Statechart.Build.MacroChart
+  alias Statechart.Build.MacroOpts
   alias Statechart.Build.MacroTransition
   alias Statechart.Build.MacroSubchart
+  alias Statechart.Build.MacroTransition
   alias Statechart.Machine
+  alias Statechart.Schema.Location
+
+  @type t(context) :: %Statechart.Machine{
+          # LATER i don't like this key name
+          statechart_module: atom(),
+          context: context,
+          current_local_id: local_id(),
+          last_event_status: :ok | :error
+        }
+
+  @opaque local_id :: Location.local_id()
+  @opaque t :: t(term)
+  @type event :: term()
+  @type state :: atom()
+  @type action :: (context() -> context()) | (() -> :ok)
+  @type context :: term
 
   defmacro __using__(_opts) do
     quote do
-      import Statechart, only: [statechart: 1, statechart: 2]
+      import Statechart,
+        only: [
+          statechart: 0,
+          statechart: 1,
+          statechart: 2,
+          subchart_new: 1,
+          subchart_new: 2
+        ]
+
+      require MacroChart
+      require AccStep
     end
   end
 
   @doc section: :build
   @doc """
+  Create a statechart node.
+
+  Examples
+
+  arity-1 (name only)
+
+      statechart do
+        state :my_only_state
+      end
+
+  arity-2 (name and opts)
+
+      statechart do
+        state :state_with_opts, entry: fn -> IO.inspect "hello!" end
+                                exit: fn -> IO.inspect "bye" end
+      end
+
+  arity-2 (name and do block)
+
+      statechart do
+        state :parent_state do
+          state :child_state
+        end
+      end
+
+
+  arity-3 (name and opts and do-block)
+      statechart do
+        state :parent_state,
+          entry: fn -> IO.inspect("hello!") end,
+          exit: fn -> IO.inspect("bye") end do
+          state :child_state
+        end
+      end
+
+  module's statechart.
+  The way to have multiple nodes sharing the same name is to define statechart
+  partials in separate module and then insert those partials into a parent statechart.
+
+  #{MacroOpts.docs(:state)}
+  """
+  @spec state(state(), Keyword.t(), term()) :: term
+  defmacro state(name, opts, do_block)
+  defmacro state(name, opts, do: block), do: MacroState.build_ast(name, opts, block)
+
+  @doc """
+  Create a statechart node.
+
+  See `state/3` for details
+  """
+  @doc section: :build
+  @spec state(state(), Keyword.t() | term()) :: term
+  defmacro state(name, opts_or_do_block \\ [])
+  defmacro state(name, do: block), do: MacroState.build_ast(name, [], block)
+  defmacro state(name, opts), do: MacroState.build_ast(name, opts, nil)
+
+  @doc section: :build
+  @doc """
   Create and register a statechart to this module.
-  May only be used once per module.
 
   ```
   defmodule ToggleStatechart do
@@ -33,89 +118,23 @@ defmodule Statechart do
     end
   end
   ```
-
-  `statechart/2` accepts a `:module` option.
-  In the below example,
-  the module containing the statechart is `Toggle.Statechart`
-  ```
-  defmodule Toggle do
-    use Statechart
-
-    statechart module: Statechart do
-      state :on, default: true, do: :TOGGLE >>> :off
-      state :off, do: :TOGGLE >>> :on
-    end
-  end
-  ```
-
-  In this way, many statecharts may be declared easily in one file:
-  ```
-  defmodule MyApp.Statechart do
-    use Statechart
-
-    # module: MyApp.Statechart.Toggle
-    statechart module: Toggle do
-      state :on, default: true, do: :TOGGLE >>> :off
-      state :off, do: :TOGGLE >>> :on
-    end
-
-    # module: MyApp.Statechart.Switch
-    statechart module: Switch do
-      state :on, default: true, do: :SWITCH_OFF >>> :off
-      state :off, do: :SWITCH_ON >>> :on
-    end
-  end
-  ```
-
-  ## `StatechartError` raised when...
-  - `statechart/2` is used more than once per module
+  #{MacroOpts.docs(:statechart)}
   """
-  defmacro statechart(opts \\ [], do_block)
-
-  defmacro statechart(opts, do: block) do
-    ast = MacroStatechart.build_ast(block, Keyword.put(opts, :include_public_api, true))
-
-    case opts[:module] do
-      nil ->
-        quote do
-          (fn -> unquote(ast) end).()
-        end
-
-      module ->
-        quote do
-          defmodule unquote(module) do
-            unquote(ast)
-          end
-        end
-    end
-  end
-
-  @doc false
-  defmacro state(name) do
-    MacroState.build_ast(name, [], nil)
-  end
 
   @doc section: :build
+  defmacro statechart(opts, do_block)
+  defmacro statechart(opts, do: block), do: MacroChart.build_ast(:statechart, opts, block)
+
   @doc """
-  Create a statechart node.
+  Create or register a statechart to this module.
 
-  `name` must be an atom and must be unique amongst nodes defined in this
-  module's statechart.
-  The way to have multiple nodes sharing the same name is to define statechart
-  partials in separate module and then insert those partials into a parent statechart.
-
-  ## `StatechartError` raised when...
-  - `name` is non-atom
-  - `name` is non-unique (another node already has the same name)
-  - assigning a default to a leaf node
-  - a default targets a non-descendent
-  - `state/2` is called outside of a `statechart` block
+  See `statechart/2` for details.
   """
-  defmacro state(name, opts \\ [], do_block)
+  @doc section: :build
+  defmacro statechart(opts_or_do_block \\ [])
 
-  defmacro state(name, opts, do: block) do
-    MacroState.build_ast(name, opts, block)
-  end
+  defmacro statechart(do: block), do: MacroChart.build_ast(:statechart, [], block)
+  defmacro statechart(opts), do: MacroChart.build_ast(:statechart, opts, nil)
 
   # FutureFeature
   @doc false
@@ -123,135 +142,78 @@ defmodule Statechart do
   # ## `StatechartError` raised when...
   # - `subchart/2` is passed anything besides the name of a module that containing a `statechart/2` call
   # - `state/2` is called outside of a `statechart` block
-  defmacro subchart(name, module, do_block \\ [do: nil])
+  defmacro subchart(name, module, opts \\ [], do_block \\ [do: nil])
 
-  defmacro subchart(name, module, do: block) do
-    MacroSubchart.build_ast(name, module, block)
+  defmacro subchart(name, module, opts, do: block) do
+    MacroSubchart.build_ast(name, module, opts, block)
   end
 
+  # LATER rename to subchart, add doc, and make public
+  # have to then remove the current subchart and absorb its functionality into `state`
+  _doc = """
+  blarg
+
+  blarg
+
+  #{MacroOpts.docs(:subchart)}
+  """
+
+  @doc false
+  defmacro subchart_new(), do: MacroChart.build_ast(:subchart, [], nil)
+  @doc false
+  defmacro subchart_new(do: block), do: MacroChart.build_ast(:subchart, [], block)
+
+  defmacro subchart_new(opts) do
+    MacroChart.build_ast(:subchart, opts, nil)
+  end
+
+  @doc false
+  defmacro subchart_new(opts, block), do: MacroChart.build_ast(:subchart, opts, block)
   @doc section: :build
   @doc """
   Register a transtion from an event and target state.
-
-  ## `StatechartError` raised when...
-  - `event` is non-atom
-  - `event` occurs elsewhere amongst this node's ancestors or descendents
-  - `target_state` doesn't exist
-  - `>>>/2` is called outside of a `state` block
   """
   defmacro event >>> target_state do
     MacroTransition.build_ast(event, target_state)
   end
 
-  @doc false
-  _future_doc = """
-  Declare an action (zero-arity function) to be run when a node is entered or exited.
-
-  There are two available actions:
-  ```
-  on enter: &do_something_when_entering_a_node/0
-  on exit: &do_something_else_when_exiting/0
-  ```
-
-  ## Add action to statechart
-  This can be used at the top level of a `Statechart.statechart/2`...
-  ```
-  defmodule ToggleStatechart do
-    use Statechart
-
-    statechart do
-      on enter: fn -> IO.puts "ToggleStatechart.machine/0 was just called" end
-      state :on, default: true, do: :TOGGLE >>> :off
-      state :off, do: :TOGGLE >>> :on
-      on exit: fn -> IO.puts "This will never print" end
-    end
-  end
-  ```
-
-  ## Add action to state node
-
-  More often though, it is used inside of a state:
-  ```
-  defmodule ToggleStatechart do
-    use Statechart
-
-    statechart do
-      state :on, default: true do
-        on enter: fn -> IO.puts "Turn on" end
-        :TOGGLE >>> :off
-        on exit: fn -> IO.puts "Exit :on" end
-      end
-
-      state :off do
-        on enter: fn -> IO.puts "Turn off" end
-        :TOGGLE >>> :on
-        on exit: fn -> IO.puts "Exit :off" end
-      end
-    end
-  end
-  ```
-
-  These actions are then triggered as follows
-  ```
-  ToggleStatechart.machine()
-  # "Turn on"
-  |> Statechart.Machine.transition(:TOGGLE)
-  # "Exit :on"
-  # "Turn off"
-  |> Statechart.Machine.transition(:TOGGLE)
-  # "Exit :off"
-  # "Turn on"
-  |> Statechart.Machine.transition(:TOGGLE)
-  # "Exit :on"
-  # "Turn off"
-  ```
-
-  ## `StatechartError` raised when
-  - it is passed anything other than a keyword list with a single key-value pair
-  - the key is anything other than `t:Statechart.action_type/0`
-  - `on2` is called outside of a `state` block
+  @doc section: :manipulate
+  @doc """
+  Get current context data.
   """
-
-  defmacro on(action_and_function)
-
-  defmacro on([{action_type, action_fn}]) do
-    MacroOnEnterExit.build_ast(action_type, action_fn)
-  end
+  @spec context(t(context)) :: context when context: var
+  defdelegate context(statechart), to: Machine
 
   @doc false
   defmacro root() do
     quote do: __MODULE__
   end
 
-  @opaque statechart :: Statechart.Machine.t()
-  @type event :: term()
-  @type state :: term()
-
   @doc section: :manipulate
   @doc """
   Send an event to the statechart
   """
-  @spec trigger(statechart, event) :: statechart
+  @spec trigger(t(context), event) :: t(context) when context: var
   defdelegate trigger(statechart, event), to: Machine
-
-  @doc section: :manipulate
-  @doc """
-  Get the current compound state
-  """
-  @spec states(statechart) :: [state]
-  defdelegate states(statechart), to: Machine, as: :states
 
   @doc section: :manipulate
   @doc """
   Determine if the given state is in the given compound state
   """
-  @spec in_state?(statechart, state) :: boolean
+  @spec in_state?(t, state) :: boolean
   defdelegate in_state?(statechart, state), to: Machine
 
   @doc section: :manipulate
   @doc """
-  Returns `:ok` is last event was valid and caused a transition
+  Returns `:ok` if last event was valid and caused a transition
   """
-  @spec last_event_status(statechart) :: :ok | :error
+  @spec last_event_status(t) :: :ok | :error
   defdelegate last_event_status(statechart), to: Machine
+
+  @doc section: :manipulate
+  @doc """
+  Get the current compound state
+  """
+  @spec states(t) :: [state]
+  defdelegate states(statechart), to: Machine, as: :states
 end