+++ lib/statechart/util/devdoc.ex
@@ -1,17 +1,10 @@
 defmodule Statechart.Util.DevOnlyDocs do
   @moduledoc false
   @render? System.get_env("RENDERDEVDOCS") == "true"
-  # @render? true
 
   defmacro __using__([{:moduledoc, moduledoc}]) do
-    if @render? do
-      quote do
-        @moduledoc unquote(moduledoc)
-      end
-    else
-      quote do
-        @moduledoc false
-      end
+    quote do
+      @moduledoc unquote(@render?) && unquote(moduledoc)
     end
   end
 end